WIP refactor paths
