+++ src/repositories/user.ts
@@ -1,5 +1,5 @@
 import { Logger } from '@nestjs/common';
-import { User, UserRepository } from '../types/user.types';
+import { User, UserRepository } from '../types/user';
 import fetch from 'node-fetch';
 
 export class RestUserRepository implements UserRepository {
